# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API de Gerenciamento de Projetos",
      "version": "1.0.0",
      "description": "Documentação da API de Gerenciamento de Projetos"
    },
    "servers": [
      {
        "url": "http://localhost:8081",
        "description": "Rota base para a API"
      }
    ],
...
    "paths": {
      "/api/pessoas": {
        "get": {
          "tags": [
            "Pessoas"
          ],
          "summary": "Busca todas as pessoas cadastradas (Necessita privilégio de administrador).",
          "description": "Retorna lista completa de pessoas cadastradas (Necessita privilégio de administrador).",
          "operationId": "getPessoas",
          "responses": {
            "200": {
              "description": "Ok",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Pessoa"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Token de autenticação não informado ou inválido"
            },
            "403": {
              "description": "Acesso não autorizado"
            }
          },
          "security": [
            {
              "token_acesso": []
            }
          ]
        }
      },
      "/api/pessoa/{Id}": {
        "get": {
          "tags": [
            "Pessoas"
          ],
          "summary": "Busca pessoa pelo Id",
          "description": "Retorna uma única pessoa (Necessita privilégio de administrador).",
          "operationId": "getPessoaById",
          "parameters": [
            {
              "name": "Id",
              "in": "path",
              "description": "ID da pessoa a ser retornada.",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Pessoa"
                  }
                }
              }
            },
            "401": {
              "description": "Token de autenticação não informado ou inválido"
            },
            "403": {
              "description": "Acesso não autorizado"
            },
            "404": {
              "description": "Pessoa não encontrada"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          },
          "security": [
            {
              "token_acesso": []
            }
          ]
        }
      },
      "/api/cadastrarPessoa": {
        "post": {
          "tags": [
            "Pessoas"
          ],
          "summary": "Cria um cadastro de pessoa",
          "description": "Cria um cadastro de pessoa (Necessita privilégio de administrador).",
          "operationId": "postPessoa",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Criado",
              "content": {
                "application/json": {
                  "schema": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Pessoa"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "token": {
                            "type": "string",
                            "example": "zI1NiI5cCI6IkpXVCJ9.IjoiQ0FRElEQVRpYXQiOjE2ODU5MTg1OTI1OTcwLCM2ODU5NzB9.bhqomndFXsNPPEdtE"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            },
            "401": {
              "description": "Token de autenticação não informado ou inválido"
            },
            "403": {
              "description": "Acesso não autorizado"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          },
          "security": [
            {
              "token_acesso": []
            }
          ]
        }
      },
      "/api/editarPessoa/{Id}": {
        "put": {
          "tags": [
            "Pessoas"
          ],
          "summary": "Atualiza um cadastro de pessoa",
          "description": "Atualiza os dados de um cadastro de pessoa pelo Id (Necessita privilégio de administrador).",
          "operationId": "putPessoa",
          "parameters": [
            {
              "name": "Id",
              "in": "path",
              "description": "ID da pessoa a ser atualizada.",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Ok",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Pessoa"
                  }
                }
              }
            },
            "401": {
              "description": "Token de autenticação não informado ou inválido"
            },
            "403": {
              "description": "Acesso não autorizado"
            },
            "404": {
              "description": "Pessoa não encontrada"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          },
          "security": [
            {
              "token_acesso": []
            }
          ]
        }
      },
      "/api/deletarPessoa/{Id}": {
        "delete": {
          "tags": [
            "Pessoas"
          ],
          "summary": "Elimina pessoa pelo Id",
          "description": "Elimina um registro de pessoa pelo Id (Necessita privilégio de administrador).",
          "operationId": "deletePessoa",
          "parameters": [
            {
              "name": "Id",
              "in": "path",
              "description": "ID da pessoa a ser eliminada.",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Eliminado"
            },
            "401": {
              "description": "Token de autenticação não informado ou inválido"
            },
            "403": {
              "description": "Acesso não autorizado"
            },
            "404": {
              "description": "Pessoa não encontrada"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          },
          "security": [
            {
              "token_acesso": []
            }
          ]
        }
      },
...
      "/api/candidatos": {
        "get": {
          "tags": [
            "Candidatos"
          ],
          "summary": "Busca todas os candidatos cadastrados.",
          "description": "Retorna lista completa de candidatos cadastrados (Necessita privilégio de administrador).",
          "operationId": "getCandidatos",
          "responses": {
            "200": {
              "description": "Ok",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Pessoa"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Token de autenticação não informado ou inválido"
            },
            "403": {
              "description": "Acesso não autorizado"
            }
          },
          "security": [
            {
              "token_acesso": []
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
        "Pessoa": {
          "required": [
            "nome",
            "idade",
            "cpf",
            "email",
            "role"
          ],
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "readOnly": true,
              "example": "647d30526ba74d90b5a905bf"
            },
            "nome": {
              "type": "string",
              "example": "Candidato 1"
            },
            "idade": {
              "type": "integer",
              "example": 20
            },
            "cpf": {
              "type": "string",
              "example": 99999999999
            },
            "email": {
              "type": "string",
              "example": "cand@email.com"
            },
            "role": {
              "type": "string",
              "enum": [
                "ADMIN",
                "RESPONSAVEL",
                "CANDIDATO"
              ],
              "example": "CANDIDATO"
            }
          }
        },
...
      },
      "securitySchemes": {
        "token_acesso": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }